Fix wrong release tag in dependent packages
While looping over the list of packages, we manipulated the package
release and changelog, which affected all subsequent packages. Instead
of modifying the values passed to the function, copy them and only
modify the copy. This way, in the next loop iteration, we're using the
original values passed to the function.