copr_build: move fetching builds into separate function, cache results
This speeds up the build process because it only fetches the list of
builds once, for subsequent calls, the cached results are used. Note
that this only works if you build the same package only once during one
run, as the list of builds is not updated after the first fetch.