Add option to only generate and build new packages with --only-new
Sometimes we want to add a package and all its dependencies to the
repository without updating any existing dependencies (i.e., it should
use the version in the repository for the dependency). For this use
case, add an option '--only-new', which only generates SPEC files for
non-existing packages and only build packages that are not in the
repository yet.