build tree: also wait for runtime dependencies
We also need to wait for runtime dependencies as a third package may
need the package during build, where we also need the first dependency.
Consider the case where:
A requires B
C build requires A

If we build A before B (which was possible because it is not a build
requirement) and then try to build C, installing A will fail, because B
has not been built. Thus, wait for B before building A.