build tree: add state SKIPPED for packages that should not be built
This is useful if we non-recursively build packages. The dependencies of
a package will also be added to the tree, but should not be built. Mark
those as SKIPPED so we can immediately continue with the actual package
to be built.

Also mark the tree as built even if it contains skipped packages.