rosfed: allow multiple system packages with the same name
When we resolve the name of a package to a system package name, we
expect exactly one result. However, the system may resolve the name to
multiple packages with the same name. Store the result in a set and
check that the set has length 1 instead of requiring a list with length
1.