rosfed.py: update default distro to noetic
The current default distro is kinetic, but that's already EOL according
to the ROS wiki (http://wiki.ros.org/Distributions). So let's update to
noetic instead which is the latest ROS release.

Also, while being there let's print the default distro in the rosfed.py
help text and also have a section in the README that mentions the distro.

Fixes: #5